fix: base component for modals
