+++ lib/components/Modal/base/modal.scss
@@ -1,0 +1,113 @@
+@import "@quill/breakpoints.scss";
+.quill-modal {
+    &__background {
+        position: fixed;
+        left: 0;
+        right: 0;
+        top: 0;
+        bottom: 0;
+        z-index: 9999;
+        display: flex;
+        align-items: center;
+        background-color: var(--core-color-opacity-black-600);
+    }
+    &__container {
+        &--visible {
+            opacity: var(--core-opacity-1300);
+            transform: translateY(0%);
+            pointer-events: auto;
+        }
+
+        &--expanded {
+            min-height: 90%;
+            max-height: 90%;
+        }  
+    }
+
+    &__handle-bar {
+        position: absolute;
+        cursor: pointer;
+        top: var(--semantic-spacing-general-sm);
+        left: calc(50% - (var(--core-size-2400) / 2));
+        width: var(--core-size-2400);
+        height: var(--core-size-200);
+        background-color: var(--core-color-opacity-black-300);
+        border-radius: var(--core-borderRadius-100);
+
+        // Pseudo-element for increasing swipeable area
+        &:before {
+            content: "";
+            height: var(--core-size-1500);
+            width: 100vw;
+            // Half of the screen and half of the handle-bar width
+            left: calc((-100vw / 2) + (var(--core-size-2400) / 2));
+            top: calc(var(--core-size-50) - var(--semantic-spacing-general-sm));
+            position: absolute;
+        }
+    }
+    &__close-icon {
+        position: absolute;
+        cursor: pointer;
+        top: var(--core-spacing-1600);
+        right: var(--core-spacing-1600);
+    }
+    &__content {
+        &-wrapper {
+            flex-grow: 1;
+            overflow-y: scroll;
+        }
+
+        &-image {
+            min-height: var(--temp-static-spacing-144);
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            &-size {
+                height: 202px;
+                @include breakpoint("lg") {
+                    height: 288px;
+                }
+            }
+           
+
+            & + .quill-modal-bottom__content-title {
+                margin-top: var(--semantic-spacing-general-lg);
+                @include breakpoint("lg") {
+                    margin-top:  var(--core-spacing-1200);
+                }
+            }
+        }
+
+        &-title {
+            margin: var(--core-spacing-2200) var(--semantic-spacing-general-lg)
+                0;
+                @include breakpoint("lg") {
+                    margin: var(--core-spacing-1600);
+                }
+        }
+
+        &-body {
+            margin: var(--semantic-spacing-general-lg)
+                var(--semantic-spacing-general-lg);
+                @include breakpoint("lg") {
+                    margin: var(--core-spacing-1200) var(--core-spacing-1600);
+                }
+        }
+    }
+    &__button {
+        border-width: var(--core-borderWidth-75);
+
+        &-wrapper {
+            align-self: stretch;
+            padding: var(--semantic-spacing-general-lg);
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            flex-direction: column;
+            gap: var(--semantic-spacing-gap-md);
+            @include breakpoint("lg") {
+                padding: var(--core-spacing-1200) var(--core-spacing-1600);
+            }
+        }
+    }
+}
